fix warnings
